:truck: move  `CertificationContract` to `certification/evaluation`
